fix all (most?) interactive commands to not modify mark in place.
The mark is indicating a significant buffer position.  An interactive command
should not use `set-mark` to directly change this "significant" buffer
position.  Instead, it should `push-mark`.

I do some acrobatics to actually use `set-mark` because we are calling many
functions in `er/try-expand-list`, and if they each pushed a mark, then the
`mark-ring` fills up with garbage.

The net effect of this commit is (hopefully), all interactive commands push
the initial buffer position onto the mark-ring, and does not take up
additional mark-ring spots.