Loosen dependency between core and mode expansions.
Extract the functionality of expand-region.el into
expand-region-core.el. The expansions depend upon that,
and expand-region.el pulls it all together.

While the dependency feature-wise is not circular
(all the mode-specific expansions depend on the core,
and not the other way around), expand-region.el should
not load all of its expansions, making it impossible
to cherry-pick what you want.

Ref issue #24