Tree-sitter support for Emacs 29+
er-basic-expansions.el (er/mark-ts-node): New function for expanding
the region using Tree-sitter parse tree nodes when Tree-sitter is
available and can parse the buffer around point. The new function is
safe to add to the global base value of `er/try-expand-list` since it
leaves the region unchanged when Tree-sitter information is not
available.